requirements.txt
django>=5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.28.0
twilio>=9.0.0
whitenoise>=6.5.0
